+++ CHANGELOG.md
@@ -2,6 +2,7 @@ v0.5.0 (in development)
 -----------------------
 - Fix "Saving torrent to file" message when torrent is actually being written
   to stdout
+- Added support for configuration files
 
 v0.4.0 (2025-05-17)
 -------------------